slight cleanup.
